+++ galleria/plugins/__init__.py
@@ -1,0 +1,5 @@
+"""Galleria plugin system."""
+
+from .base import BasePlugin
+
+__all__ = ["BasePlugin"]